refactor(mondo): overview page model
